DTSAC3Source: исправлено переполнение при поиске на больших файлах.
git-svn-id: svn://mpc-next.ru/mpc-be/trunk@4970 d191f3aa-8897-45be-b9a4-56e5fb472b67

+++ src/filters/source/DTSAC3Source/DTSAC3Source.cpp
@@ -416,7 +416,7 @@ HRESULT CDTSAC3Stream::FillBuffer(IMediaSample* pSample, int nFrame, BYTE* pOut,
 	UNREFERENCED_PARAMETER(subtype);
 
 	if (*majortype == MEDIATYPE_Audio) {
-		m_file.Seek(m_dataStart + nFrame*m_framesize, CFile::begin);
+		m_file.Seek(m_dataStart + (__int64)nFrame * m_framesize, CFile::begin);
 		if ((int)m_file.Read(pOut, m_framesize) < m_framesize) {
 			return S_FALSE;
 		}
